requirements.txt
streamlit>=1.30.0
openai>=1.12.0
groq>=0.4.0
google-generativeai>=0.4.0
duckduckgo-search>=4.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
